% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addExamHistory.R
\name{addExamHistory}
\alias{addExamHistory}
\title{Function to add comment for questions that have been part of an exam}
\usage{
addExamHistory(exam_list, exam_name)
}
\arguments{
\item{exam_list}{list.}

\item{exam_name}{character. Name of the exam}
}
\description{
Function to add comment for questions that have been part of an exam
}
\examples{
addExamHistory("Kapitalwert_01_S0_mc.Rnw", "just a test")
}
