% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addTable.R
\name{addTable}
\alias{addTable}
\title{Wrapper function for xtable
used for simple tables like bond characteristics}
\usage{
addTable(dat, rotate = TRUE, hafter = NULL)
}
\arguments{
\item{dat}{dataframe.}

\item{rotate}{logical.}
}
\value{
latex code with content of dataframe
}
\description{
Wrapper function for xtable
used for simple tables like bond characteristics
}
