% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Anleihe.R
\docType{class}
\name{Anleihe-class}
\alias{Anleihe-class}
\title{An S4 class to represent a bond}
\description{
An S4 class to represent a bond
}
\section{Slots}{

\describe{
\item{\code{nom}}{numeric. Nominale}

\item{\code{emk}}{numeric. Emissionskurs}

\item{\code{i}}{numeric. Kupon}

\item{\code{t}}{numeric. Laufzeit}

\item{\code{typ}}{character. valid is "serie", "zero", "float", "end"}

\item{\code{fee}}{numeric. Nachschuessige Geb}

\item{\code{initfee}}{numeric Einmalige Geb}

\item{\code{tilgung}}{numeric vector. Tilgungsplan}

\item{\code{zinsen}}{numeric vector. Zahlungsstrom Zinsen}

\item{\code{schuld}}{numeric vector. Restschuld}

\item{\code{fees}}{numeric vector. Zahlungsstrom Geb}

\item{\code{zahlungen}}{numeric vector. Zahlungsstrom Summe}
}}

\section{Constructor}{

\code{newAnleihe(nom=100, ...)}
}

\section{Methods}{

\itemize{
\item \code{getAnTyp(an)} gives a string representation (e.g. "Zerobond")
\item \code{getAnTable(an)} returns a latex table with cash flows
\item \code{getAnKonditionen(an)} returns a latex table with all characteristics
\item \code{getAnWert(an, rlz, i)} returns the NPV based on maturity and interest rate
\item \code{getAnClean(an, rlz, i)} returns the fair clean price based on maturity and interest rate
\item \code{getAnDirty(an, rlz, i)} returns the fair dirty price based on maturity and interest rate
\item \code{getAnStueckzinsen(an, rlz, i)} returns the accrued interest based on maturity
}
}

\examples{
an <- newAnleihe(100, 0.05, 4) ## a standard bond from issuers perspective
an@zinsen ## returns c(0, -5, -5, -5, -5)
an@zahlungen ## returns c(100, -5, -5, -5, -105)
getAnWert(an, 2.75, 0.04) ## returns -103.7878 (issuers perspective!)

}
