% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Asset.R
\docType{class}
\name{Asset-class}
\alias{Asset-class}
\title{An S4 class to represent an asset}
\description{
Used in some tasks to scramble stuff that you own. Challenge is always to calculate the NPV.
}
\section{Slots}{

\describe{
\item{\code{txt}}{character. Verbal representation}

\item{\code{k0}}{numeric. NPV}

\item{\code{distractor}}{numeric. Distractor, can be identical to k0}
}}

\section{Constructor}{

\code{newAsset(type, ieff, nmax, kmax, n, k)}
type is:
\itemize{
\item 1: Zukuenftige Zahlung
\item 2: Rente, Start in t0 oder t1
\item 3: Zahlung in der Vergangenheit
\item 4: Rente in der Vergangenheit
\item 5: Ein Barwert
}
nmax integer. Maximale Laufzeit
kmax numeric. Maximaler NPV
}

\section{Methods}{

\itemize{
\item \code{getAssetTxt(ass)} getter for ass@txt
\item \code{getVal(ass)} getter for ass@k0
\item \code{getDistractor(ass)} getter for ass@distractor
}
}

\examples{
newAsset(2, 0.05, 6, 100000)

}
