% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkIfEqual.R
\name{checkIfEqual}
\alias{checkIfEqual}
\title{Assert function to make sure that two numbers are different.
If the two numbers are too similar, execution will stop with error
It will use catg and tnum (standard in each iuf question) in the error message to identify the problematic question}
\usage{
checkIfEqual(k1, k2, prec = 0)
}
\arguments{
\item{k1}{numeric.}

\item{k2}{numeric.}

\item{prec}{integer. Presision to check}
}
\description{
Assert function to make sure that two numbers are different.
If the two numbers are too similar, execution will stop with error
It will use catg and tnum (standard in each iuf question) in the error message to identify the problematic question
}
\examples{
checkIfEqual(1.1155, 1.1166, prec=2) ## will stop with error as rounding to 2 digits is identical
checkIfEqual(1.1155, 1.1166, prec=3) ## will pass though
}
