% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkIfGreater.R
\name{checkIfGreater}
\alias{checkIfGreater}
\title{Assert function to make sure that second parameter is strictly greater than first parameter.
If not, execution will stop with error
It will use catg and tnum (standard in each iuf question) in the error message to identify the problematic question}
\usage{
checkIfGreater(k1, k2, prec = 0)
}
\arguments{
\item{k1}{numeric.}

\item{k2}{numeric.}

\item{prec}{integer. Presision to check}
}
\description{
Assert function to make sure that second parameter is strictly greater than first parameter.
If not, execution will stop with error
It will use catg and tnum (standard in each iuf question) in the error message to identify the problematic question
}
\examples{
checkIfGreater(1.6, 1.5, prec=2) ## will stop with error as 2nd parameter is smaller
checkIfGreater(1.1155, 1.1166, prec=2) ## will stop with error as rounding to 2 digits is identical
checkIfGreater(1.1155, 1.1166, prec=3) ## will pass though
}
