% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Darlehen.R
\docType{class}
\name{Darlehen-class}
\alias{Darlehen-class}
\title{An S4 class to represent a loan}
\description{
An S4 class to represent a loan
}
\section{Slots}{

\describe{
\item{\code{k0}}{numeric. Nominale}

\item{\code{i}}{numeric. Darlehenszins}

\item{\code{t}}{numeric. Laufzeit}

\item{\code{typ}}{character. valid is "ann", "end", "const"}

\item{\code{fee}}{numeric. Nachschuessige Geb}

\item{\code{initfee}}{numeric Einmalige Geb}

\item{\code{freijahre}}{integer.}

\item{\code{tilgung}}{numeric vector. Tilgungsplan}

\item{\code{zinsen}}{numeric vector. Zahlungsstrom Zinsen}

\item{\code{schuld}}{numeric vector. Restschuld}

\item{\code{fees}}{numeric vector. Zahlungsstrom Geb}

\item{\code{zahlungen}}{numeric vector. Zahlungsstrom Summe}
}}

\section{Constructor}{

\code{newDarlehen(typ="??", nom, i, t, ...)}
}

\section{Methods}{

\itemize{
\item \code{getTyp(dar)} gives a string representation (e.g. "Zerobond")
\item \code{getTable(dar)} returns a latex table with cash flows
\item \code{getKonditionen(dar)} returns a latex table with all characteristics
}
}

\examples{
dar <- newDarlehen(typ="ann", 1000, 0.05, 4) ## Annuitaetendarlehen aus Sicht Schuldner
dar@zinsen ## returns c(0.00000, -50.00000, -38.39941, -26.21879, -13.42913)
dar@zahlungen ## returns c(1000.0000, -282.0118, -282.0118, -282.0118, -282.0118)

}
