% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eform.R
\name{eform}
\alias{eform}
\title{Number to String.
This function converts a number to a string with . as thousand sep , as decimal sep and it rounds the number with a hand-made rounding funtion. It should make sure that float imprecision problems are mitigated
Thus, do not use eform(round(1235.55, 1), 1)! Let eform do rounding!}
\usage{
eform(n, digits = -1)
}
\arguments{
\item{n}{numeric. The number to convert}

\item{digits}{numeric. Number of digits, -1 for displaying all significant digits}
}
\value{
character. Representation of the number
}
\description{
Number to String.
This function converts a number to a string with . as thousand sep , as decimal sep and it rounds the number with a hand-made rounding funtion. It should make sure that float imprecision problems are mitigated
Thus, do not use eform(round(1235.55, 1), 1)! Let eform do rounding!
}
\examples{
eform(1235.55, 1)

}
