% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatChoices.R
\name{formatChoices}
\alias{formatChoices}
\title{Function to convert a raw_choices vector to a formatted choices vector}
\usage{
formatChoices(
  raw_choices,
  prefix = "",
  suffix = "",
  separator = " ",
  digits = 2
)
}
\arguments{
\item{raw_choices}{character list.}

\item{prefix}{character.}

\item{suffix}{character.}

\item{separator}{character.}

\item{digits}{integer. Number of digits for formatting}
}
\description{
Function to convert a raw_choices vector to a formatted choices vector
}
