% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatString.R
\name{formatString}
\alias{formatString}
\title{Function that provices named string formatting}
\usage{
formatString(fmt, ...)
}
\arguments{
\item{fmt}{character. String containing variables}

\item{...}{character. Replacements for variables}
}
\description{
Function that provices named string formatting
}
\examples{
formatString("The \%{coffee}s costs \%{price}s EUR.", coffee="cappuccino", price=2.8)
## The cappuccino costs 2.8 EUR
}
