% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fwform.R
\name{fwform}
\alias{fwform}
\title{Names and formats a forward rate}
\usage{
fwform(i, n2 = 1, n1 = 0, slash = 2)
}
\arguments{
\item{i}{numeric. Interest rate}

\item{n2}{integer. End of period}

\item{n1}{integer. Start of period}

\item{slash}{integer. Numer of backslashes for formatting}
}
\description{
Names and formats a forward rate
}
