% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fwform.R
\name{fwname}
\alias{fwname}
\title{Names a forward rate}
\usage{
fwname(n2 = 1, n1 = 0, slash = 2)
}
\arguments{
\item{n2}{integer. End of period}

\item{n1}{integer. Start of period}

\item{slash}{integer. Numer of backslashes for formatting}
}
\description{
Names a forward rate
}
