% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDifficulty.R
\name{getDifficulty}
\alias{getDifficulty}
\title{Function that translates difficulty levels [0;3] to points [1;3]
Used primarily in OLAT examples pool and Onlinetests. For written exams the points for each questions are provided separately.}
\usage{
getDifficulty(diff)
}
\arguments{
\item{diff}{integer. Between 0 and 3, is specified in each iuf exam question!}
}
\value{
integer. Points that OLAT assignes for correct answer of that question
}
\description{
Function that translates difficulty levels [0;3] to points [1;3]
Used primarily in OLAT examples pool and Onlinetests. For written exams the points for each questions are provided separately.
}
\examples{
usually at the bottom of an iuf question the metainformation uses getDifficulty(diff) to pass diff to specify points
}
