% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIntervalPool.R
\name{getIntervalPool}
\alias{getIntervalPool}
\title{Get pool questions and solutions for intervals (possibly) containing the correct result}
\usage{
getIntervalPool(correct_result, max_dev = 0.2, digits = 2)
}
\arguments{
\item{correct_result}{numeric. The correct result is the reference point}

\item{max_dev}{numeric. Maximum positive and negative deviation of intervals from correct result}

\item{digits}{integer.}
}
\value{
list. List containing the question pool and solution pool
}
\description{
Get pool questions and solutions for intervals (possibly) containing the correct result
}
\examples{
getIntervalPool(1000, 10, 2)

}
