% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRandomDistractors.R
\name{getRandomDistractors}
\alias{getRandomDistractors}
\title{Add a number of random distractors
drawn from a range vector while making sure that the random distractors do not coincide with existing choices in the exclude vector}
\usage{
getRandomDistractors(
  num_distractors = 4,
  range = c(0, 1),
  digits = 2,
  exclude = c()
)
}
\arguments{
\item{num_distractors}{integer. Number of distractors to produce}

\item{range}{numeric vector length 2. Range for distractors}

\item{digits}{integer.}

\item{exclude}{numeric vector.}
}
\description{
Add a number of random distractors
drawn from a range vector while making sure that the random distractors do not coincide with existing choices in the exclude vector
}
