% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmlPreview.R
\name{htmlPreview}
\alias{htmlPreview}
\title{Wrapper function for exams2html}
\usage{
htmlPreview(questions, edir = "Questions", template = "plain.html")
}
\arguments{
\item{questions}{character list. Questions}

\item{edir}{character. Directory with questions}

\item{template}{character. Template to use}
}
\description{
Wrapper function for exams2html
}
