% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ieff.R
\name{ieff}
\alias{ieff}
\title{calculate effective compounding factor}
\usage{
ieff(i, m)
}
\arguments{
\item{i}{numeric. interest rate}

\item{m}{integer. interest periods per year. 0 for continous compounding}
}
\description{
calculate effective compounding factor
}
