% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeBsp.R
\name{makeBsp}
\alias{makeBsp}
\title{Wrapper for exams2openolat to create unscrambeled pool
Sets the global variable MAKEBSP to TRUE (to set seed in each exam) and creates 1 scrambling}
\usage{
makeBsp(name, myexam)
}
\arguments{
\item{name}{character. Exam name}

\item{myexam}{list. Exam files to include}
}
\description{
Wrapper for exams2openolat to create unscrambeled pool
Sets the global variable MAKEBSP to TRUE (to set seed in each exam) and creates 1 scrambling
}
