% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeBsp.R
\name{makeBspHTML}
\alias{makeBspHTML}
\title{Wrapper for exams2html
Sets the global variable MAKEBSP to TRUE (to set seed in each exam) and creates 1 scrambling}
\usage{
makeBspHTML(name, myexams)
}
\arguments{
\item{name}{character. Exam name}

\item{myexams}{list. Exam files to include}
}
\description{
Wrapper for exams2html
Sets the global variable MAKEBSP to TRUE (to set seed in each exam) and creates 1 scrambling
}
