% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeBsp.R
\name{makeBspPDF}
\alias{makeBspPDF}
\title{Wrapper for exams2pdf
Sets the global variable MAKEBSP to TRUE (to set seed in each exam) and creates 1 scrambling}
\usage{
makeBspPDF(name, myexams)
}
\arguments{
\item{name}{character. Exam name}

\item{myexams}{list. Exam files to include}
}
\description{
Wrapper for exams2pdf
Sets the global variable MAKEBSP to TRUE (to set seed in each exam) and creates 1 scrambling
}
