% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeOT.R
\name{makeOT}
\alias{makeOT}
\title{Wrapper for exams2openolat to create online test with 100 scramblings}
\usage{
makeOT(name, myexam)
}
\arguments{
\item{name}{character. Exam name}

\item{myexam}{list. Exam files to include}
}
\description{
Wrapper for exams2openolat to create online test with 100 scramblings
}
