% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSK.R
\name{makeSK}
\alias{makeSK}
\title{Wrapper for exams2openolat to create corona online final exam with 100 scramblings}
\usage{
makeSK(name, myexam)
}
\arguments{
\item{name}{character. Exam name}

\item{myexam}{list. Exam files to include}
}
\description{
Wrapper for exams2openolat to create corona online final exam with 100 scramblings
}
