% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSt.R
\name{makeSt}
\alias{makeSt}
\title{Wrapper for exams2openolat to create scrambeled example pool with 20 scrambling}
\usage{
makeSt(name, myexam)
}
\arguments{
\item{name}{character. Exam name}

\item{myexam}{list. Exam files to include}
}
\description{
Wrapper for exams2openolat to create scrambeled example pool with 20 scrambling
}
