% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbw.R
\name{rbw}
\alias{rbw}
\title{Calculate NPV of annuity}
\usage{
rbw(k, q, n)
}
\arguments{
\item{k}{numeric. annuity}

\item{q}{numeric. discounting factor}

\item{n}{integer. periods}
}
\value{
NPV
}
\description{
Calculate NPV of annuity
}
