% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbwfg.R
\name{rbwfg}
\alias{rbwfg}
\title{Calculate annuity factor for growing annuity}
\usage{
rbwfg(i, n, g = 0, q = 1 + i)
}
\arguments{
\item{i}{numeric. interest rate}

\item{n}{integer. periods}

\item{g}{numeric. growth rate}

\item{q}{numeric. discounting factor}
}
\description{
Calculate annuity factor for growing annuity
}
