% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig.R
\name{sig}
\alias{sig}
\title{Create standfardized signature for a iuf exams question}
\usage{
sig(t, prec = 0, diff = -1)
}
\arguments{
\item{t}{character. The type of quest: "mc" (multiple choice), "mco" (multiple choice at least one correct), "sc" (single choice), "num" (one numeric result), "per" (one percent result), "pers" (several percent results), "perp" (percentage points), "perps" (several percentage points)}

\item{prec}{integer. The precision, i.e. number of digits to round}

\item{diff}{integer. Depreciated}
}
\value{
string with the complete signature
}
\description{
Create standfardized signature for a iuf exams question
}
\examples{
sig("mc", prec = 2)
}
