% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whatToDo.R
\name{whatToDo}
\alias{whatToDo}
\title{Function to show all todos in a directory or an exam list}
\usage{
whatToDo(dir = "Questions", exam_list = NULL)
}
\arguments{
\item{dir}{character. Directory relative to working dir (must be iuf!)}

\item{exam_list}{list. Optional, can contain exam questions (only filenames)}
}
\description{
Function to show all todos in a directory or an exam list
}
\examples{
whatToDo()
whatToDo("Kapitalwert_01_S0_mc.Rnw")
}
